+++ packages/types-core/CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2023-10-31
+### Added
+- New `endBlock` option on datasources (#2064)
+
 ## [0.2.0] - 2023-10-20
 ### Added
 - Add cache to injected globals (#2110)
@@ -21,7 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release of new package containing types common to all chains
 
-[Unreleased]: https://github.com/subquery/subql/compare/types-core/0.2.0...HEAD
+[Unreleased]: https://github.com/subquery/subql/compare/types-core/0.3.0...HEAD
+[0.3.0]: https://github.com/subquery/subql/compare/types-core/0.2.0...types-core/0.3.0
 [0.2.0]: https://github.com/subquery/subql/compare/types-core/0.1.0...types-core/0.2.0
 [0.1.0]: https://github.com/subquery/subql/compare/types-core/0.0.1...types-core/0.1.0
 [0.0.1]: https://github.com/subquery/subql/tag/types-core/0.0.1